chore: break down classes in files
